requirements.txt
discord.py==1.3.4
numpy==1.18.4
Pillow==7.1.0
pyttanko==1.2.1
requests==2.23.0
SQLAlchemy==1.3.17
